Replace BitSet with a custom implementation.
Now uses a set that will only grow as the number of open frames grows
instead of growing as the frame id increases. Also uses an immutable
style set that making copying unnecessary and no longer need to guard
set reads. Copies only now need to be done when setting and clearing
a bit in the set which is performed by the set and clear calls.

Test: new tests and existing tests pass

Change-Id: Ic25b8b1541345d99b32f6c4fa98107c8eaf52d3e