Hide the fingerprint dialog on certain devices
The biometric library currently has a list of devices for which we show
FingerprintDialogFragment as a workaround but which have a conflicting
UI (such as an overlay) that requires the dialog to be dismissed
immediately upon receiving an error. While this is an improvement, it
would arguably be better for the fingerprint dialog not to show at all
on these affected devices. This commit makes that change, using the
preexisting list of affected device models.

Test: Biometric demo app on Galaxy S10 and Note10+
Fixes: 143230260
Change-Id: I81a43db82126cd7f594f2328e3ca9b0158c33881