Make GuidedStepFragment touch screen friendly
Fixed EditText bug that it "consumed" TouchEvent and prevented
parent from receiving onClick() events.

The fix logic is:
If EditText is not focusable, not editable and not text selectable,
we should propagate the TouchEvent to parent.
If EditText is focusable or editable or text selectable, then it
probably need handle TouchEvent.

Also removed container view focusableInTouchMode=true, it will incorrectly
let the container view gain focus on first touch. By setting
focusableInTouchMode  to false, first touch will translated to click
which is the common user experience on a touchscreen.

GuidedStepFragment is still not touch screen friendly for editing text, and there
is a bigger problem that IME is not touch screen friendly.

Bug: 143562736
Test: SupportLeanbackDemos

Change-Id: I632a112462538155507908f98cb2cd7af0ee3be5