Deprecate setRetainInstance() method in Fragment
SetRetainInstance() seeks to maintain the state of a fragment across
activity re-creation. It comes with the side effect of a much more
complicated lifecycle combinations when mixed with Views, non-retained
fragments, and or setTargetFragment().

Using ViewModels to retain the state of a fragment accomplishes the same
goal, but because the fragment with state stored in a view model is not
special in anyway, it has none of the mentioned side effects.
Therefore, it is now better to retain fragment state using a view model.

With the introduction of ViewModels, there is no longer a need for
setRetainInstance() and we should deprecate it, and promote the use of
ViewModels instead.

Test: ./gradlew checkApi
BUG: 143911815
Change-Id: I2cd277180bd8fff78c4b23d87641f3c3422bb945