Keep FragmentManager state consistent with Transactions
Previously, if a FragmentTransaction executed a remove operation, once
the transaction was completed, the FragmentManager state was updated
immediately, regardless of whether there were any animations. Changes in
Fragments 1.2 made it so that onDestroyView() is not called until
animations/transitions are complete, which means the FragmentManager
state must wait for the animation/transition to finish before it can
settle.

The state of the FragmentManager should be settled once a
FragmentTransaction completes, with or without animations/transitions.

This change ensures that once a FragmentTransaction that executes a
remove operation completes, a fragment with an animation  being
removed is no longer recognized by the FragmentManager, even though
that fragment will maintain a reference to the FragmentManager until
it finishes its exit animation, and is destroyed.

Test: Added Tests
BUG: 143982969
Change-Id: I7a5be24866689f6675fbf51ed84593e9afd8f612