Dispatch onTouch to behaviors as onTouch.
Partially reverts a previous change to the touch dipsatch that changed
the dispatch in onTouchEvent to send onInterceptTouchEvent to the
behaviors instead of onTouchEvent. By not sending an onTouchEvent
behaviors do not get a chance to decide that they wish to handle touches
that are otherwise unhandled by descendants so this is not a safe
change.

Test: ./gradlew coordinatorlayout:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.package=androidx.coordinatorlayout.widget
Change-Id: I81d1c4c909f5c4f39ee8edc29e88068491ddbccc