Fix cache clearing regression in text benchmarks.
In aosp/1109940 I have moved some work from setupContentInternal to
local field. This means that it was called only once per test and not
every-time content was composed. This erased the original purpouse it
had to clear caches.

This fix changes the lifespan of the test case. So from now on the test
case does not survive more than one call of "emitContent" this ensures
that any work done in the initialization of the test case is related to
just that one composition. This makes it easier to reason about and also
makes the previous change valid.

The benchmarks are now slower again. But locally I don't have that
significant difference as in the benchmarks. I have several other ideas
of what changed and will look into it (e.g. composition disposal now
works properly which did not work before).

Bug: 144086917
Test: Tried the benchmarks locally
Change-Id: I8c8ef010041f1b8f007671ca7e25d692c2b439c1