Make FingerprintHelperFragmentTest a local unit test
Modifies FingerprintHelperFragmentTest so it can be run as a local unit
test instead of as an instrumented android test and moves it from
androidTest to the test directory. By running as a local unit test, it
can now mock otherwise final parts of the Android framework, such as
various Handler methods. This makes the MessageRouter pass-through class
introduced for testing unnecessary, so this commit also removes that.

Test: ./gradlew biometric:test
Test: ./gradlew biometric:connectedAndroidTest

Bug: 139874508
Change-Id: Ic3975c0e6500f5b649be89d797b7f1d404fc8f9b