Finite measurement for inflexible Flex children
The CL updates the layout algorithm of Row and Column to measure
inflexible children with finite constraints. Before this CL, these were
measured with infinite main axis constraints to obtain their preferred
size. This had the advantage that "fill available space" children layout
models would be unable to fill the whole Row/Column. However, this has
proven to come at a high cost in the predictability of the behavior of
Row/Column, which is not worth the advantage.

Bug: 143165169
Bug: 137168151
Test: added new tests in FlexTest
Change-Id: I793bb365f2221a5456c10f8f4339d18b2f695a09