Animation Clock
This CL introduced the concept of an animation clock observable. All animations
are observers of an animation clock source. And they will subscribe to the
data source in the duration of the animation. The default clock can be
read out from the ambient and is pulsed with Choreographer frame callbacks.

Devs can supply their own custom the animation clock and overwrite the ambient
clock. By doing this, all the animations in the subtree from the point of
the ambient clock overwrite will subscribe to that custom clock. The benefits for
such a design: 1) Animations can be deterministically tested as devs manipulate
the animation clock at a very fine grain. 2) Android Studio can also drive a custom
clock to see the frame-by-frame preview of the animation

Also provided ManualAnimationClock class out of the box to support manually
driving the clock.

Test: included in the CL

Change-Id: I5f61027e4ccf6dda658b244f8d724c758c2f96d8