Adding FixedSizeSurfaceTexture as temporary fix in camera-view
The TextureViewImplementation uses FixedSizeSurfaceTexture to prevent
TextureView from modifying the SurfaceTexture size, since the
SurfaceTexture needs to be the exact size specified by
PreviewSurfaceProvider.provideSurface().

Removed PreviewViewTest.defaultImplementation_isSurfaceView because the
PreivewView will be choosing the best implementation for the device
automatically which won't always be a SurfaceView.

Added workaround for SurfaceTextureReleaseBlockingListener test which
caused native crash in API 28 and below (see b/144878737 for details).
However, the crash has not been seen to exhibit it self in app usage.

Bug: b/144386349
Test: Manual test on view-testapp
Change-Id: I572ce065db72103bb28c8b46317bec4f3aa61831