Fix hide Transitions with reordering
It should be possible to replace a replace transaction with transitions
with a combination of add and hide while seeing the same results. That
is currently impossble. While the initial transactions works, if you
attempt to pop the transactions it breaks because all non-shared
elements are invisible.

This is caused because when using a hide with transitions, any exitings
views are marked as invisible during the transitions and are not
properly marked visible again with the transition ends.

This change ensures that once the transition ends, all exiting views are
properly marked as visible so that a hide transaction with transitions
can be applied and popped properly.

Test: All tests pass, tested in several apps with and without shared
elements, adjusted FragmentTransitionTest
BUG: 70793925

Change-Id: Idb4991b898c966910ebc56ae155e10b64222e819