onSizeChange event when modifiers' size changes
Before this CL, we would only let the owner know about size changes when
the modified size of a LayoutNode changed. However, this meant that if
the modified size of the LayoutNode did not change, but the size of the
LayoutNode or the size of an intermediate wrapper changed, the owner
would not find out. This would cause repaint boundaries to not be
invalidated correctly. This CL fixes this by tracking size changes for
intermediate wrappers as well.

Bug: 145730617
Test: ran AndroidLayoutDrawTest\#layoutModifier_redrawsCorrectlyWhenOnlyNonModifiedSizeChanges
Change-Id: I995765c194a51f6e409df22140cdedf7657158ad