Allow lazy state restoration driven by the Adapter
It is a common problem for layouts to lose scroll position due to a
configuraiton change when Adapter loads its data asyncronously.

So far, the only solution was to avoid setting the adapter on the
RecyclerView, which was not easy to implement + didn't allow adapter
to show any other items(like loading spinners).

This CL adds a state restoration mode to the adapter which has 3
options:

IMMEDIATE: Current behavior, also the default to avoid any issues;
RecyclerView restores state immediately.
WHEN_NOT_EMPTY: RecyclerView restores state as soon as Adapter has 1
or more items.
MANUAL: RecyclerView does not restore state until another call to
setStateRestorationMode is made with another value.

As state restoration happens in the LayoutManager, this change affects
how LayoutManager restores state. Now it receives the restored state
only when it can use it. Until then, state is kept in RecyclerView.

I've also changed SavedStateTests to add an index to each test
run so that it is easy to re-run a failing test.

Bug:146365793
Test: LinearLayoutManagerSavedStateTest, SGLMSavedStateTest,
LazyStateRestorationTest
Change-Id: I35004de2f63dcb3428a6f5d8ca54d5c8cdfa0798