Allow lazy state restoration driven by the Adapter
It is a common problem for layouts to lose scroll position due to a
configuraiton change when Adapter loads its data asyncronously.

So far, the only solution was to avoid setting the adapter on the
RecyclerView, which was not easy to implement + didn't allow adapter
to show any other items(like loading spinners).

This CL adds a state restoration mode to the adapter which has 3
options:

IMMEDIATE: Current behavior, also the default to avoid any issues;
RecyclerView restores state immediately.
WHEN_NOT_EMPTY: RecyclerView restores state as soon as Adapter has 1
or more items.
MANUAL: RecyclerView does not restore state until another call to
setStateRestorationMode is made with another value.

As state restoration happens in the LayoutManager, this change affects
how LayoutManager restores state. Now it receives the restored state
only when it can use it. Until then, state is kept in RecyclerView.

I've also changed SavedStateTests to add an index to each test
run so that it is easy to re-run a failing test.

Note: This is a re-implementation of a previously reverted Change
I35004de2f63dcb3428a6f5d8ca54d5c8cdfa0798 which broke a common use
case where devs called LayoutManager#onRestore multiple times.
That method is not intended to be called outside RV but as it is
public API, such behaviors are not surprising.
This version slightly changes how LayoutManager ignores saved state.
In the previous implementation, it had a long living flag which
made state restoration impossible after a scrollToPosition call.
This implementation instead ignores it if and only if there is a
pending scroll position.

I would still prefer the previous implementation as Views are not
designed to restore state multiple times yet there is existing
code that relies on that behavior and unfortunately LayoutManager#
onRestoreInstance state is public (due to previous API guidelines of
banning protected) so this solution finds a middle ground in between.

Bug:146365793
Bug:148104407
Test: LinearLayoutManagerSavedStateTest, SGLMSavedStateTest,
LazyStateRestorationTest, GridLayoutManagerSavedStateTest

Change-Id: I43050128fcd3842190a1bb3fa98ffd724844f0ef