Change FragmentFinishEarlyTest to use StrictFragment
The goal of the FragmentFinishEarlyTest is to ensure that the Fragment
state is not raised while the activity is being destroyed.
StrictFragment now checks if the activity is being destroyed in all of
the upward states (onCreate, onStart, etc.) so we should just use that
instead of a custom Fragment with booleans.

Test: all tests pass
BUG: 144309266
Change-Id: Ic47a178c50bcbba246b91e2e25d969dd6931d765