Solve chaining of injected gestures
This solves the problem at the InputDispatcher level, not yet at the
GestureScope level. It might not be necessary to solve the problem at
the GestureScope level though, if we provide methods for all gestures
that require chaining.

Solved by adding state to AndroidInputDispatcher, storing at what time
the next gesture should start, and adding a delay command to
InputDispatcher. This allows gestures to be chained by performing them
sequentially on the same InputDispatcher, optionally adding a delay
between them, without the need to modify existing API.

Bug: 145593518
Test: ./gradlew ui:ui-test:cC
Change-Id: Ic9e759fe0f07dcdb0ff13e839f533771b3e517c0