Un-finalize FingerprintManagerCompat class
The FingerprintManagerCompat class is currently final, which developers
have called out as making it difficult to test dependent classes such as
BiometricPrompt. This class was made final in 2015, and the justification
at the time seems to have been a bit vague (see b/24461189). Given the
unclear benefits and definite drawbacks, this commit updates the class
to be non-final.

Test: ./gradlew buildOnServer
Bug: 142545354
Change-Id: I6f9b4b530afbf3dacf06854553bb28fdbd1bd041