Add "require___()" lint check for Fragments
This adds a new lint to suggest require___() APIs in place of checkNotNull(get____()) calls, as the former will yield more descriptive error messages if it fails. Works on both Kotlin and Java code.

We've been using this in the Slack codebase for a few weeks without issues, and figured it would be worthwhile for upstream consideration!

Full checks:

- `checkNotNull`
- `requireNonNull`
- `!!`

Note that only single-argument nullcheck methods are supported. If a message is defined then the lint will skip it and leave the message in tact.

Test: Unit tests
Change-Id: I32a9fc1fba3b78ed76f80b8b6214ca1ae44bfacb