Fix test crash by java.util.concurrent.RejectedExecutionException
Under real circumstances, the executor is provided by the developer,
developer should take responsibility to the lifetime of the executor.
The solution is to use the main thread executor instead of custom
executor, then we don't have to care the lifetime of the executor.
If the callback is not triggered, the result should be verified through
Truth or Mockito APIs, rather than crashing the test.
The CL also extends the timeout to pass the test.

Bug: 147272013, 147401932
Test: ./gradlew camera:camera-camera2:connectedAndroidTest
Change-Id: I1c4d51599574a7f80f5208f16395e17e51aa189e