AndroidX Webkit: IDs for medium-sized WebViews
No change to library, only testapp.

This adds unique IDs for the medium-sized WebViews used to demonstrate
"quiet" Safe Browsing interstitials. This declares the WebViews in the
layout instead of programmatically creating them and appending them to
the layout. This should be simpler logic long-term, at the cost of
needing to update 2 parts of the app instead of 1 when we add new threat
types.

Bug: 147497976
Test: Manual - app still works for manual use, but should now have
Test: unique IDs
Change-Id: Ie124b66aa4e14fed96d7be28769250e581cd8c2f