Fix crash when Text and WithConstraints are siblings
This was happening because WithConstraints is calling FrameManager.nextFrame() after subcomposition inside the measure block and Text is changing the @Model annotated state value inside it's measure block. This is triggering requestRemeasure call on the Text right during the measuring of the Text and WithConstraints parent. Now we handle such case by postponing requestRemeasure till the end of the measure/layout pass if we detect we can't apply it right now. I rewrote our remeasuring/relayout logic to make it more predictable and stable and covered with new assertions. Plus introduced an additional opt-in assertions which traverse the tree, print it's state and asserts the consistency of the tree, makes sure all the contracts are still in place. I propose to only enable this additional assertion during our core layout tests like AndroidLayoutDrawTest and WithConstraintsTest. This should not affect the performance when the assertions are not enabled.

Bug: 147219075
Test: tested manually, added new tests, old tests are passing with new assertions
Change-Id: I8ff98099312dfb35575301754c0be4d1b45b8f56