Use compiler schema for enum coder generation
Use the schema implementation defined in
androidx.serialization.compiler.schema for enum coder generation instead
of the schema interfaces defined in androidx.serialization.schema. This
is less relevant for enums, but will be relevant for messages and
services.

To enable testing code generators in isolation, this CL adds
implementations for a subset of the interfaces in
javax.lang.model.element. These are consumed by scaffolding functions to
produce instances of the compiler schema suitable for use with code
generators.

Test: ./gradlew :serialization:serialization-compiler:test
Bug: 144724751
Change-Id: I83128e53fb627e6b9b8ef8f2be3d47c44effc697