Ensure query params are ignored for all paths in deeplinks
When adding a deeplink without any query params and trying to match with
a deeplink that has query params, matching fails. If we get a path with
no query params, we should ignore all of them.

This change makes sure the query params are the end of the path are
always optional.

Test: Added NavDeepLinkTest
Bug: 147447512
Change-Id: I46256bed27f9384526c0f38258e817f5f26a1b5d