Remove unused CameraUtil method for opening camera
The openCameraWithUseCase should not be used. Instead for now
CameraX.bindToLifecycle() should be used for testing. Eventually
CameraX.bindToLifecycle() should also be removed in favor of a simpler
method for binding Camera and UseCase together. This change is to reduce
the number of places where variations of binding a Camera and UseCase
exist.

Also cleaned up some lint suppression in coretestapp.

Bug: b/147698557
Test: ./gradlew bOS
Change-Id: I9016bd4b4043d537b266f88459c0775bffa4bc13