Move Release Notes Script to python
This CL refactors the buildSrc version of the release-notes
automation and tests into a python scrip that can
be run without gradle.  Given that all release information is
retrieved from Jetpad, there is no reason to keep it
in the build.

Additionally, this let's us use this script in branches
that don't have the buildSrc changes.

Test: ./testReleaseNotes.py
Test: ./generateReleaseNotes.py 1581346800000
Change-Id: Ieac9e4f069fab44c9f362462cb0c648e7d526545