Fix FragmentContainerView errors in layout preview
When using FragmentContainerView in xml and going to the layout preview
an UnsupportedOperationException will be thrown because of the
contructor that the layout preview uses to initialize the view.

This change adds a isInEditMode() check to avoid incorrectly spamming
the editor with errors.

Test: Tested in nav integration app, and test app
Bug: 149707833
RelNote: "`FragmentContainerView` no longer throws an
`UnsupportedOperationException` for using the incorrect constructor in
layout preview."

Change-Id: Ia9850eaeb1ad5dad654c62441b9ba0827e30663e