Base class for all method "deprecation" calls + getDrawable() rule
The new base class is a predicate-based system that allows
"flagging" a number of existing APIs along with custom conditions
such as SDK level. When all conditions associated with a specific
flagged API call are satisfied, the call is reported to Lint.

The existing ColorStateListLoadingDetector has been switched to
the new system. A new DrawableLoadingDetector which is based on
the new system has been added as well.

Bug: 147878022
Test: ./gradlew :lint-demo:lint-demo-appcompat:lint
Test: ./gradlew :appcompat:appcompat-lint:test
Change-Id: I9dd6528883147cb22991569668082f51f2f25885