Added UiSavedStateRegistry
This is a first component which will be used by upcoming Saved Instance State support mechanism for Compose UI.
The contract of this class is pretty similar to the one we have in androidx.savedstate.SavedStateRegistry, so I decided it's reasonable to have the same name, but add Ui prefix to emphasize it's a Compose UI specific implementation.
New ui-saved-instance-state module is added. In the future it will also contain functions replacements for state/remember which the will be used by end users. I decided to go with a separate module as this Registry will be used by ui-platform, so we can theoretically move it to ui-core, but in the same time this module will contain @Composable function savedInstanceState, and ui-core doesn't have any composables yet

Bug: 151084498
Test: covered with UiSavedStateRegistryTest
Change-Id: Ib5d88c6cf6e41e71aaafbff51e20c765efa08fd3