Add MeteringRepeating and MeteringRepearingConfig
In current design, when using ImageCapture with CaptureMode.MAXIMIZE_QUALITY,
it must bind another repeating use case (Preview or ImageAnalysis).
Even for CAPTURE_MODE_MINIMIZE_LATENCY, it need to do auto 3A check,
or the image quality is poor w/o the repeating use case. The goal
is make ImageCapture takePicture w/o the repeating usecase and could
resolve the above issues.

Design Doc: go/camerax-takepicture-no-repeating-usecase

Use a MeteringRepeating UseCase to be the alternative of Preview or
ImageAnalysis when ImageCapture only. Add/Remove the metering use case
when tryAddOnLineUseCases or tryRemoveOnLineUseCases.

The 1st chain CL is to add MeteringRepeating UseCase and MeteringRepeatingConfig

Bug: 145326998
Test: ./gradlew bOS && ./gradlew camera:camera-*:connectedCheck
Change-Id: I4ffe382e1535df9885351f563c65a09337de43f4