Fix rarely happening layout tree inconsistency
Sometimes the LayoutNodes added or invalidated during the subcomposition were not marked as needsRemeasure due to incorrect assumption that if the parent is measuring and (layout.measureIteration != measureIteration) then the child will be measured soon. It will not happen if the child has no needsRemeasure set to true. They were in fact remeasured not because of the needsRemeasure, but because of the changed constraints. New logic is simpler and handles this case, plus layout consistency checker is updated to be aware about the new logic.

Test: tests are passing, new test
Change-Id: I1701075900a09d939d62152f4179c69056d6c02e