Ensure we makeInactive when we destroy fragments
After we destroy the fragment, we never call makeInactive. This creates
an issue where if the activity is recreated, the fragment store still
contains references to old fragments. If a user attempts to re-add that
old fragment, an IllegalStateException is thrown because the
childFragmentManager is reattached to its host already.

We should make sure we call FragmentStore.makeInactive when we destroy
the Fragment.

Test: add test in FragmentLifecycleTest
Fixes: 151237982
Change-Id: I99a3f4eeced4922677e12cfbbba53d59764f146e