Allow only siblings to rearrange Z-order drawing.
This CL creates a layer when necessary to prevent
improper ordering of drawn elements. When a layout
contains a child whose outer layer has an elevation,
a layer is necessary to ensure that children of
the layout don't paint with Z order with cousins.
This CL adds a layer in that case only when one
hasn't already been introduced.

Fixes: 151398183
Test: new tests. ran demo
Change-Id: I44f099a483eca6b6630fb914d55941790f23704b