Adds `generateDocs` flag to allow opting out for docs generation
This allows keeping API tracking turned on, but disabling docs module-wide
for projects that are published but not intended to be used directly, yet are used
by other libraries.

Additionally generateDocs is false by default for any unpublished libraries, so demos / testapps
no longer need to be opted-out of docs.

Bug: 142788501
Test: ./gradlew dokkaKotlinTipOfTreeDocs, see that demos are not generated there
Change-Id: Ia05725314828ae108310577951be57fed3b8b867