Make dialog null when mShowsDialog is false
Since setShowsDialog() is a public API, mShowsDialog can be changed at
any point in the lifecycle of the dialog fragment. If mShowsDialog is
made false in onCreateDialog(), at which we currently had already
decided we would show the dialog, then we can get into a state where
mShowsDialog is false and we have a non-null dialog that will be shown
anyway.

We should ensure that up until we actually show the dialog in onStart()
that mShowsDialog and whether the dialog is null are in sync.

Test: added test in DialogFragmentTest
Bug: 152414644
Change-Id: I7fec4b22876d1e18043f0a5b1a5c1e62687a08ab